requirements.txt
# ===== Core scientific stack =====
numpy
scipy
pandas
matplotlib
tqdm
PyYAML
ipykernel

# ===== PyTorch (CPU wheels) =====
torch